Add deletion by occurrence and modify rewrite
Now it is possible to complete a to-do composed by multiple words with
"-c <word>". For example, if my to-do is "2) buy some flowers", i can
complete it with "-c flowers". If multiple todos contains that word
(phrase), all those todos are listed.

The rewrite function it is rewritten so it doesn't take an integer. The
problem is that now the list passed must reversed, otherwise every time
a deletion occurs the whole list is rewritten in reversed order.